Saw trap working, started on another trap
I have created a saw trap that works and damages the player, it moves up and down on the Y axis, and can be configured to move at any speed, distance wanted. You can also change the damage it does. I have also started to implement a moving block trap, but this is not working yet.