Death Menu Started
I have started working on the death menu, the death menu currently uses the background that Ravn created with the text elements that he also created. Currently all the buttons work, but there is no background so the default blue is shown. I made a new script to control these buttons and it works well but the replay button is NOT CURRENTLY ASSIGNED TO ANYTHING USEFULL.