Death Menu complete with new fade animation.
On death the screen slowly fades to black. This is done through a raw image component in the UI which starts 100% see through, and then fades to 0% see through when the death is triggered. When the player health drops below zero, the GameController is called to execute the death function, which calls the "prepareFade" function from the DeathFade script attached to the rawimage, and then starts the fade proccess itself. This could be made more efficient as the DeathFade script is attached to the DeathFade object, it would not be nececary for this object to get itself.