Dialogue box added in simple test form.
I have added the dialogue box function, but currently its just a feature in testing, and is manually triggered using the space key. This is not going to be suitable for the main game, but its simple to make it useable in other ways as the function is simply called from any other script in the game. I am planning to use simple triggers to execute the script.