All dialogues for Level 1 added.
All the dialogues and their triggers for Level 1 are now implemented and they all work as the player progresses along the level. I had to edit Julian's Dialogue nr2 because we do not have the feature that lets you interact with walls, all the dialogues are dynamic, we don't have time to implement interactable elements.