Pruning for release
I have removed all the copyrighted material we used for testing of the game, and all old folders that aren't in use. This is deliberately done on a separate branch because this is only for the first release candidate. Development will continue on imptest. I have also removed all the old test scenes. The evaluation should still be done based on the imptest branch as most of our work was testing.